40/requirements.txt
discord.py==1.7.3
pandas==1.3.4
